Update 4 digit & segment display
 7-segment display is a form of electronic display device used to represent decimal numerals and other characters. The display is composed of seven individual segments, each of which can be individually activated to display the desired numeral or character.


Segment Pins and Decimal Point Control


Out of the 12 pins on the display, the segments are controlled by the pins without resistors, while one pin is dedicated to controlling the decimal point. These segment pins allow for the activation of specific segments to form the desired numeral or character on the display.


Digit Pins


The remaining four pins, each equipped with a resistor, are responsible for controlling each of the four digits on the display. By activating specific combinations of segments through these pins, different numerals or characters can be displayed on each of the four digits.


Functionality


By selectively activating the appropriate segments through the segment pins and utilizing the digit pins to control each digit, the 7-segment display can effectively showcase a wide range of numerals and characters, making it a versatile and widely used component in various electronic applications.